Stepper: remove plans from critical path (#109302)
