+++ client/landing/stepper/declarative-flow/internals/steps-repository/unified-plans/util/get-visual-split-plans-intent.ts
@@ -1,0 +1,11 @@
+import { type PlansIntent } from '@automattic/plans-grid-next';
+
+export function getVisualSplitPlansIntent( intent?: string | null ): PlansIntent | null {
+	if ( intent === 'default_websitebuilder' ) {
+		return 'plans-website-builder';
+	}
+	if ( intent === 'default_hosting' ) {
+		return 'plans-wordpress-hosting';
+	}
+	return null;
+}
