+++ client/landing/stepper/declarative-flow/internals/steps-repository/unified-plans/util/package.json
@@ -1,0 +1,3 @@
+{
+	"sideEffects": false
+}
