+++ client/landing/stepper/declarative-flow/internals/steps-repository/unified-plans/util/supported-interval-types.ts
@@ -1,0 +1,13 @@
+import { type UrlFriendlyTermType } from '@automattic/calypso-products';
+
+export type SupportedIntervalTypes = Extract<
+	UrlFriendlyTermType,
+	'monthly' | 'yearly' | '2yearly' | '3yearly'
+>;
+
+export const supportedIntervalTypes: SupportedIntervalTypes[] = [
+	'monthly',
+	'yearly',
+	'2yearly',
+	'3yearly',
+];
